% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcor_test_cpp}
\alias{pcor_test_cpp}
\title{Projection Correlation Permutation Test for Independence}
\usage{
pcor_test_cpp(X, Y, estimation_method = "u", times = 199L, n_threads = 4L)
}
\arguments{
\item{X}{A numeric matrix of dimension n x p, where each row is an i.i.d. observation from the first vector.}

\item{Y}{A numeric matrix of dimension n x q, where each row is an i.i.d. observation from the second vector.}

\item{estimation_method}{A character string, either \code{"u"} or \code{"v"}.
If \code{"u"}, U-statistics are used for estimation; if \code{"v"}, V-statistics are used. Default is \code{"u"}.}

\item{times}{An integer specifying the number of permutations to perform. Default is 199.}

\item{n_threads}{An integer specifying the number of threads to use for parallel computation. Default is 4.}
}
\value{
A list with components:
\item{method}{The method name: \code{"Projection Correlation Permutation Test of Independence"}}
\item{stat.value}{The test statistic value (scaled by sample size).}
\item{p.value}{The p-value computed from the permutation distribution under the null hypothesis of independence.}
}
\description{
Perform a permutation test to assess the independence between two random vectors
via projection correlation. The two vectors can have different dimensions but
must share the same sample size.
}
\examples{
X <- matrix(rnorm(10 * 7, 1), 10, 7)
Y <- matrix(rnorm(10 * 6, 1), 10, 6)
pcor_test_cpp(X, Y, estimation_method = "v", times = 199, n_threads = 4)

}
\references{
L. Zhu, K. Xu, R. Li, W. Zhong (2017). Projection correlation between two random vectors. Biometrika, 104, 829-843. \doi{10.1093/biomet/asx043}
}
\seealso{
\code{\link{pcov_cpp}}, \code{\link{pcor_cpp}}
}
